backend/requirements.txt
# Core Flask
Flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Security
bcrypt>=4.0.0
PyJWT>=2.8.0

# HTTP Requests (for AI APIs)
requests>=2.31.0
httpx>=0.24.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# AI Libraries
google-generativeai>=0.3.0
groq>=0.4.0

# Email
Flask-Mail>=0.9.1

# File Handling
filelock>=3.13.0

# Utilities
python-dateutil>=2.8.0

# Validation
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
